+++ edge_side_pkg/EPGeneral_mqtav/test/__init__.py
@@ -1,1 +1,0 @@
-"""mqtav test package."""